Update stop scripts

+++ scripts/stop_server.sh
@@ -2,4 +2,6 @@
 
 # TODO: improved server shutdown logic
 process_id=$(netstat -nlp | grep :8000 | awk '{print $NF}' | awk -F "/" '{print $1}')
-kill -9 $process_id
+if [ -n "${process_id}" ]; then
+    kill -9 $process_id
+fi
